requirements.txt
pyTelegramBotAPI
aiohttp
peewee
pyaml